requirements.txt

keras==2.9.0
Keras-Preprocessing==1.1.2
matplotlib-inline==0.1.6
numpy==1.22.3
Pillow==9.0.1
tensorflow==2.9.1
numpy==1.22
pandas==1.1.3 
streamlit==1.9.0
scikit-learn==0.23.2